@@ -3,7 +3,7 @@ package co.csadev.kwikpickerdemo
 import android.Manifest
 import android.net.Uri
 import android.os.Bundle
-import android.support.v7.app.AppCompatActivity
+import androidx.appcompat.app.AppCompatActivity
 import android.util.TypedValue
 import android.view.*
 import android.widget.*
