@@ -3,7 +3,7 @@ package co.csadev.kwikpicker.view
 import android.content.Context
 import android.graphics.Canvas
 import android.graphics.drawable.Drawable
-import android.support.v7.widget.AppCompatImageView
+import androidx.appcompat.widget.AppCompatImageView
 import android.util.AttributeSet
 
 import co.csadev.kwikpicker.R
