@@ -1,15 +1,14 @@
 package co.csadev.kwikpicker.view
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
 import android.util.AttributeSet
 import android.view.View
 
-class KwikEmptyRecyclerView(context: Context, attrs: AttributeSet? = null) : RecyclerView(context, attrs) {
+class KwikEmptyRecyclerView(context: Context, attrs: AttributeSet? = null) : androidx.recyclerview.widget.RecyclerView(context, attrs) {
     private var emptyView: View? = null
 
-    private val observer: RecyclerView.AdapterDataObserver =
-        object : RecyclerView.AdapterDataObserver() {
+    private val observer: androidx.recyclerview.widget.RecyclerView.AdapterDataObserver =
+        object : androidx.recyclerview.widget.RecyclerView.AdapterDataObserver() {
             override fun onChanged() {
                 super.onChanged()
                 checkIfEmpty()
@@ -17,10 +16,10 @@ class KwikEmptyRecyclerView(context: Context, attrs: AttributeSet? = null) : Rec
         }
 
     internal fun checkIfEmpty() {
-        emptyView?.visibility = if (adapter.itemCount > 0) View.GONE else View.VISIBLE
+        emptyView?.visibility = if ((adapter?.itemCount ?: 0) > 0) View.GONE else View.VISIBLE
     }
 
-    override fun setAdapter(adapter: RecyclerView.Adapter<*>?) {
+    override fun setAdapter(adapter: androidx.recyclerview.widget.RecyclerView.Adapter<*>?) {
         val oldAdapter = getAdapter()
         oldAdapter?.unregisterAdapterDataObserver(observer)
         super.setAdapter(adapter)
